requirements.txt
tensorflow==2.2.0
matplotlib==3.2.2
scikit-learn==0.23.1
Pillow==7.1.2
flask==1.1.2
opencv-python==4.2.0.34